az_py_upload/requirements.txt
azure-identity
azure-storage-blob